+++ src/utils/crashpad.ts
@@ -1,0 +1,5 @@
+const { crashReporter } = require('electron');
+
+export function initCrashPad() {
+    crashReporter.start({ uploadToServer: false });
+}
